+++ packages/riverpod_analyzer_utils/CHANGELOG.md
@@ -3,6 +3,7 @@
 - **Breaking**: Rewrote all RiverpodAst nodes to instead be extensions on `AstNodes`.
   Too many changes to detail everything. I'm the only one who uses this package anyway.
   If you're reading this, have a nice day!
+- Added support for parsing `@mutation`
 
 ## 1.0.0-dev.1 - 2023-11-20